% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_database}
\alias{add_database}
\title{Add Database}
\usage{
add_database(database_path, tool, database_name, ...)
}
\arguments{
\item{database_path}{Path or list of paths (local or S3) to be passed in as inputs.}

\item{tool}{Toolchest tool with which you use the database (e.g. `toolchest$tools$Kraken2``).}

\item{database_name}{Name of the new database.}
}
\description{
Adds a custom database and attaches it to a tool.
The new database version is returned immediately after initialization.
}
\details{
This executes just like any other tool, except:
\itemize{
\item the success status is \code{complete} instead of \code{ready_to_transfer_to_client}
\item \code{is_async} is True by default
}

Note that it may take 24-48 hours for the custom database to be ready for use.
}
