% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{alphafold}
\alias{alphafold}
\title{Alphafold Client}
\usage{
alphafold(
  inputs,
  output_path = NULL,
  model_preset = NULL,
  max_template_date = NULL,
  use_reduced_dbs = FALSE,
  is_prokaryote_list = NULL
)
}
\arguments{
\item{inputs}{Path or list of paths (client-side) to be passed in as input.}

\item{output_path}{(optional) Path to directory where the output file(s) will be downloaded}

\item{model_preset}{(optional) Allows you to choose a specific AlphaFold model from \link{monomer, monomer_casp14, monomer_ptm, multimer}. Default mode if not provided is monomer.}

\item{max_template_date}{(optional) Allows for predicting structure of protiens already in the database by setting a date before it was added in YYYY-MM-DD format. Will use today's date if not provided.}

\item{is_prokaryote_list}{(optional) takes a list of booleans that determine whether all input sequences in the given fasta file are prokaryotic. Expects the string that would normally input into AlphaFold (e.g. "true,true" if there are two prokaryote inputs)}

\item{use_reduced_dbs:}{(optional) Uses a smaller version of the BFD database that will reduce run time at the cost of result quality.}
}
\value{
Reference to an object with output location data.
}
\description{
Runs AlphaFold via Toolchest.
}
