% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{clustalo}
\alias{clustalo}
\title{Clustal Omega Client}
\usage{
clustalo(
  tool_args = "",
  inputs,
  output_path = NULL,
  output_primary_name = NULL,
  ...
)
}
\arguments{
\item{tool_args}{Additional arguments to be passed to Clustal Omega.}

\item{inputs}{Path (client-side) to a FASTA file that will be passed in as input.}

\item{output_path}{(optional) Path to directory where the output file(s) will be downloaded.}

\item{output_primary_name}{(optional) Base name of output file.}

\item{is_async}{(optional) Whether to run a job asynchronously. Defaults to false.}
}
\value{
Reference to an object with output location data.
}
\description{
Runs Clustal Omega via Toolchest.
}
