% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cutadapt}
\alias{cutadapt}
\title{Cutadapt Client}
\usage{
cutadapt(
  cutadapt_args,
  input_path = NULL,
  output_path = NULL,
  input_name = "input.fastq",
  output_name = "output.fastq"
)
}
\arguments{
\item{cutadapt_args}{Custom arguments given to function (e.g., adapters: "-a AATTCCGG").}

\item{input_path}{Local path of input file (to be uploaded to AWS).}

\item{output_path}{Local path of output file (to be downloaded from AWS).}

\item{input_name}{Name of input file inside input AWS S3 bucket.}

\item{output_name}{Name of output file to be uploaded to output AWS S3 bucket.}
}
\description{
Starts a query for Cutadapt using Toolchest.
}
\details{
Currently, only a single .fastq file is supported as input.
The output will be downloaded to output_path if specified.

If left unspecified, input_path and output_path can be selected by the user
manually while running the cutadapt() function.
}
\examples{
\dontrun{
cutadapt("-a AATTCCGG")
cutadapt("-a AATTCCGG", input_path = "C://Users/YourName/Documents/my_input_file.fastq")
}

}
