% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{demucs}
\alias{demucs}
\title{Demucs Client}
\usage{
demucs(tool_args = "", inputs, output_path = NULL, ...)
}
\arguments{
\item{tool_args}{Additional arguments to be passed to demucs.}

\item{inputs}{Path to a file that will be passed in as input. All formats supported by ffmpeg are allowed.}

\item{output_path}{(optional) Path where the output will be downloaded.}

\item{is_async}{(optional) Whether to run a job asynchronously. Defaults to false.}
}
\value{
Reference to an object with output location data.
}
\description{
Runs Demucs via Toolchest.
}
