% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{diamond_blastx}
\alias{diamond_blastx}
\title{Diamond BLASTx Client}
\usage{
diamond_blastx(
  tool_args = "",
  inputs,
  output_path = NULL,
  output_primary_name = NULL,
  ...
)
}
\arguments{
\item{tool_args}{Additional arguments to be passed to Diamond BLASTx.}

\item{inputs}{Path to a file that will be passed in as input. FASTA or FASTQ formats are supported (it may be gzip compressed)}

\item{output_path}{(optional) (optional) Path to directory where the output file(s) will be downloaded.
Log file (diamond.log) will be downloaded in the same directory as the out file(s).}

\item{output_primary_name}{(optional) Base name of output file.}

\item{is_async}{(optional) Whether to run a job asynchronously. Defaults to false.}
}
\value{
Reference to an object with output location data.
}
\description{
Runs Diamond in BLASTx mode via Toolchest.
}
