% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-select.R
\name{.choose.path}
\alias{.choose.path}
\title{Helper function that allows user to select a single input file path.}
\usage{
.choose.path(is_optional = FALSE, file_descriptor = "input")
}
\arguments{
\item{is_optional}{Whether choosing a file is optional.}

\item{file_descriptor}{A description of the file, to be printed when prompting the user.}
}
\value{
A file path.
}
\description{
To be used directly in place of \code{.validate.inputs()} for customization
when a certain kind of input file (e.g., long reads) is being requested.
}
