% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.do.toolchest.call}
\alias{.do.toolchest.call}
\title{Helper function that sanitizes args by removing blank args (to be handled by
the Python client) and executing the corresponding Python client call.}
\usage{
.do.toolchest.call(tool, args)
}
\arguments{
\item{tool}{Tool to be used (python function, from reticulate-based python module).}

\item{args}{List containing keyword arguments to be passed in to tool.}
}
\description{
Helper function that sanitizes args by removing blank args (to be handled by
the Python client) and executing the corresponding Python client call.
}
