% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Installs and loads the \code{toolchest_client} Python package when loading
the R package.}
\usage{
.onLoad(libname, pkgname)
}
\description{
Installs and loads the \code{toolchest_client} Python package when loading
the R package.
}
