% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-select.R
\name{.validate.inpath}
\alias{.validate.inpath}
\title{Helper function that allows user to select output path
if \code{output_file} for a tool is not specified.}
\usage{
.validate.inpath(inputs)
}
\arguments{
\item{inputs}{User-specified value for \code{inputs} variable.}
}
\value{
Input filepath(s).
}
\description{
Helper function that allows user to select output path
if \code{output_file} for a tool is not specified.
}
