% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-select.R
\name{.validate.outpath}
\alias{.validate.outpath}
\title{Helper function that allows user to select input path(s)
if \code{inputs} for a tool is not specified.}
\usage{
.validate.outpath(output_path)
}
\arguments{
\item{output_path}{User-specified value for \code{output_path} variable.}
}
\value{
Output filepath.
}
\description{
Helper function that allows user to select input path(s)
if \code{inputs} for a tool is not specified.
}
