% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download Functionality}
\usage{
download(
  output_path,
  s3_uri = NULL,
  run_id = NULL,
  pipeline_segment_instance_id = NULL,
  skip_decompression = FALSE
)
}
\arguments{
\item{output_path}{Path to a local directory where the file(s) will be downloaded.}

\item{s3_uri}{URI of file contained in S3. This can be passed from
the parameter \code{output$s3_uri} from the \code{output} returned by a previous
job.}

\item{run_id}{Run ID of the job producing the output you would like to download.}

\item{pipeline_segment_instance_id}{(DEPRECATED) Pipeline segment instance ID.
Replace this argument with \code{run_id} instead.}

\item{skip_decompression}{Whether to skip decompression of the downloaded file archive.}
}
\value{
Path(s) to downloaded and decompressed files. If \code{skip_decompression} is enabled,
returns the path to the archive.
}
\description{
If output download is skipped or halted unexpectedly during any Toolchest job,
the output file(s) can be downloaded manually via the following functions:
\itemize{
\item{\code{toolchest::download(output_path, ...)}}{As a base function}
\item{\code{output$download(output_path)}}{From the \code{output} returned by any
Toolchest job, invoking \code{download} as a function
}
}
}
\details{
If \code{toolchest::download()} is used, one of \code{s3_uri} or
\code{run_id} must be provided. These can be found as variables of the
output object returned from a previous Toolchest function call.
Contact Toolchest for assistance if the output details cannot
be recovered.

If \code{output$download()} is used from the \code{output} returned by any
\code{toolchest::tool()} function call, then only \code{output_path} needs to be specified.
}
\examples{
\dontrun{
output <- toolchest::test(...)
output$download(output_path = path)

output <- toolchest::test(...)
toolchest::download(output_path = path, s3_uri = output$s3_uri)
}

}
