% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_toolchest}
\alias{install_toolchest}
\title{Installs and loads the \code{toolchest_client} Python package.}
\usage{
install_toolchest()
}
\value{
\code{toolchest_client}, an object representing the Toolchest Python client.
}
\description{
Installs and loads the \code{toolchest_client} Python package.
}
\note{
The Python package interface is stored in the global variable \code{toolchest_client}.
}
