% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{megahit}
\alias{megahit}
\title{MEGAHIT Client}
\usage{
megahit(
  tool_args = "",
  read_one = NULL,
  read_two = NULL,
  output_path = NULL,
  interleaved = NULL,
  ...
)
}
\arguments{
\item{tool_args}{(optional) Additional arguments to be passed to MEGAHIT.}

\item{read_one}{(optional) \code{-1} inputs. Path (string) or list of paths for read 1 of paired-read input files.}

\item{read_two}{(optional) \code{-2} inputs. Path (string) or list of paths for read 2 of paired-read input files.}

\item{output_path}{(optional) Path to a directory where the output file(s) will be downloaded.}

\item{interleaved}{(optional) Path (string) or list of paths for interleaved paired-end files.}

\item{is_async}{(optional) Whether to run a job asynchronously. Defaults to false.}
}
\value{
Reference to an object with output location data.
}
\description{
Starts a query for MEGAHIT using Toolchest.
}
\details{
If \code{output_path} is left unspecified, downloading will be skipped.
}
