% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Toolchest Client Query}
\usage{
query(tool_name, tool_args, input_path, output_path, input_name, output_name)
}
\arguments{
\item{tool_name}{Toolchest tool_name to use (e.g., cutadapt).
Internal note: this is the pipeline segment name, not just the tool name.
(TODO: Resolve this naming discrepancy)}

\item{tool_args}{Custom arguments given to function.}

\item{input_path}{Local path of input file (to be uploaded to AWS).}

\item{output_path}{Local path of output file (downloaded from AWS).}

\item{input_name}{Name of input file inside input AWS S3 bucket.}

\item{output_name}{Name of output file to be uploaded to output AWS S3 bucket.}
}
\description{
Starts a generic query using Toolchest.
}
\details{
(For a specific tool, call toolchest::toolname() instead. This function
is used for processing queries and should not be called by the user itself.)
}
