% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rapsearch2}
\alias{rapsearch2}
\title{RAPSearch2 Client}
\usage{
rapsearch2(
  tool_args = "",
  inputs,
  output_path = NULL,
  output_primary_name = NULL,
  database_name = "rapsearch2_seqscreen",
  database_version = "1",
  ...
)
}
\arguments{
\item{tool_args}{(optional) Additional arguments to be passed to RAPSearch2.}

\item{inputs}{Path to a FASTA/FASTQ file that will be passed in as input.}

\item{output_path}{(optional) Path (client-side) to a directory where the output files will be downloaded.}

\item{output_primary_name}{(optional) Base name of output file(s).
(Functions the same way as the "-o" tag for RAPSearch2, in combination with \code{output_path}.)}

\item{database_name}{(optional) Name of database to use for RAPSearch2 alignment. Defaults to SeqScreen DB.}

\item{database_version}{(optional) Version of database to use for RAPSearch2 alignment. Defaults to 1.}

\item{is_async}{(optional) Whether to run a job asynchronously. Defaults to false.}
}
\value{
Reference to an object with output location data.
}
\description{
Runs Diamond in BLASTx mode via Toolchest.
}
