% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_key}
\alias{set_key}
\title{Sets the value for the Toolchest key.}
\usage{
set_key(key)
}
\arguments{
\item{key}{Value of key, or filenpath to a file containing key value.}
}
\description{
Sets the value for the Toolchest key.
}
