% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{STAR}
\alias{STAR}
\title{STAR Client}
\usage{
STAR(tool_args = "")
}
\arguments{
\item{tool_args}{(optional) Additional arguments to be passed to STAR.}

\item{inputs}{Path or list of paths (client-side) to be passed in as input.}

\item{output_path}{Path (client-side) where the output will be downloaded.}
}
\description{
Starts a query for STAR using Toolchest.
}
\details{
If left unspecified, inputs and output_path can be selected by the user
manually.
}
