% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{test}
\alias{test}
\title{Test Pipeline Segment}
\usage{
test(inputs = NULL, output_path = NULL, tool_args = "")
}
\arguments{
\item{inputs}{Path or list of paths (client-side) to be passed in as input.}

\item{output_path}{Path (client-side) where the output file will be downloaded.}

\item{tool_args}{Additional arguments, present to maintain a consistent interface. This is disregarded.}
}
\description{
Run a test pipeline segment via Toolchest. A plain text file containing 'success' is returned."
}
