% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{unicycler}
\alias{unicycler}
\title{Unicycler Client}
\usage{
unicycler(
  tool_args = "",
  read_one = NULL,
  read_two = NULL,
  long_reads = NULL,
  output_path = NULL
)
}
\arguments{
\item{tool_args}{(optional) Additional arguments to be passed to Unicycler.}

\item{read_one}{Path of input file (FASTQ) to be passed in as Read 1 (-1).}

\item{read_two}{Path of input file (FASTQ) to be passed in as Read 2 (-2).}

\item{long_reads}{Path of input file (FASTA) to be passed in as long reads (-l).}

\item{output_path}{Path (client-side) where the output file will be downloaded.}
}
\description{
Runs Unicycler via Toolchest.
}
\note{
If any of \code{read_one}, \code{read_two}, or \code{long_reads} are unspecified, Toolchest
will assume that they are intended to be blank, unlike other tools. At least
one input filepath is needed for Toolchest to run Unicycler.

The output path is needed. An option will pop up to select a filepath
if \code{output_path} is unspecified.
}
