% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{bowtie2}
\alias{bowtie2}
\title{Bowtie 2 Client}
\usage{
bowtie2(
  tool_args = "",
  inputs,
  output_path = NULL,
  database_name = "GRCh38_noalt_as",
  database_version = "1",
  ...
)
}
\arguments{
\item{tool_args}{(optional) Additional arguments to be passed to Bowtie 2.}

\item{inputs}{Path or list of paths to be passed in as input.}

\item{output_path}{(optional) Path to a directory where the output file(s) will be downloaded.}

\item{database_name}{(optional) Name of database to use for Bowtie 2 alignment. Uses the GRCh38 no-alt analysis set ("GRCh38_noalt_as") by default. Index files generated by the Langmead lab.}

\item{database_version}{(optional) Version of database to use for Bowtie 2 alignment. Defaults to "1".}

\item{is_async}{(optional) Whether to run a job asynchronously. Defaults to false.}
}
\value{
Reference to an object with output location data.
}
\description{
Starts a query for Bowtie 2 (for alignment) using Toolchest.
}
\details{
If \code{output_path} is left unspecified, downloading will be skipped.
}
