% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{bowtie2}
\alias{bowtie2}
\title{Bowtie 2 Client}
\usage{
bowtie2(
  inputs = NULL,
  output_path = NULL,
  database_name,
  database_version,
  tool_args = ""
)
}
\arguments{
\item{inputs}{Path or list of paths (client-side) to be passed in as input.}

\item{output_path}{Path (client-side) where the output file will be downloaded.}

\item{database_name}{Name (string) of database to use for Bowtie 2 alignment.}

\item{database_version}{Version (string) of database to use for Bowtie 2 alignment.}

\item{tool_args}{(optional) Additional arguments to be passed to Bowtie 2.}
}
\description{
Starts a query for Bowtie 2 (for alignment) using Toolchest.
}
\details{
If left unspecified, inputs and output_path can be selected by the user
manually.
}
