% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{clustalo}
\alias{clustalo}
\title{Clustal Omega Client}
\usage{
clustalo(tool_args = "", inputs, output_path = NULL, ...)
}
\arguments{
\item{tool_args}{Additional arguments to be passed to Clustal Omega.}

\item{inputs}{Path (client-side) to a FASTA file that will be passed in as input.}

\item{output_path}{(optional) Path (client-side) where the output file will be downloaded.}

\item{is_async}{(optional) Whether to run a job asynchronously. Defaults to false.}
}
\value{
Reference to an object with output location data.
}
\description{
Runs Clustal Omega via Toolchest.
}
