% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cutadapt}
\alias{cutadapt}
\title{Cutadapt Client}
\usage{
cutadapt(tool_args, inputs = NULL, output_path = NULL)
}
\arguments{
\item{tool_args}{Additional arguments to be passed to Cutadapt.}

\item{inputs}{Path or list of paths (client-side) to be passed in as input.}

\item{output_path}{Path (client-side) where the output file will be downloaded.}
}
\description{
Starts a query for Cutadapt using Toolchest.
}
\details{
Currently, only a single .fastq file is supported as input.
The output will be downloaded to output_path if specified.

If left unspecified, inputs and output_path can be selected by the user
manually.
}
\examples{
\dontrun{
cutadapt(tool_args = "-a AATTCCGG")
cutadapt(inputs = "C://Users/YourName/Documents/my_input_file.fastq", tool_args = "-a AATTCCGG")
}

}
