% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-select.R
\name{.validate.inpath}
\alias{.validate.inpath}
\title{Helper function that allows user to select input path(s) for a generic tool
if \code{inputs} is not specified.}
\usage{
.validate.inpath(inputs, choose_multiple = FALSE)
}
\arguments{
\item{inputs}{User-specified value for \code{inputs} variable.}

\item{choose_multiple}{Whether the prompt should be repeated to select multiple files.}
}
\value{
Input filepath(s).
}
\description{
Helper function that allows user to select input path(s) for a generic tool
if \code{inputs} is not specified.
}
