% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download}
\usage{
download(output_path, s3_uri = NULL, pipeline_segment_instance_id = NULL)
}
\arguments{
\item{output_path}{Output path to which the file(s) will be downloaded.
This should be a directory that already exists, but direct filenames
are also supported.}

\item{s3_uri}{URI of file contained in S3. This can be passed from
the parameter \code{output$s3_uri} from the \code{output} returned by a previous
job.}

\item{pipeline_segment_instance_id}{Pipeline segment instance ID of the job
producing the output you would like to download.}
}
\value{
Path(s) to downloaded and decompressed files. If \code{skip_decompression} is enabled,
returns the path to the archive.
}
\description{
Downloads the output of a previously run job.
}
\details{
One of \code{s3_uri} or \code{pipeline_segment_instance_id} must be provided.
}
\note{
The \code{download} function can also be invoked from the output of any
\code{toolchest::tool()} function call as \code{output$download(output_path)}. With an
output, only the path to the local directory needs to be specified.
}
