% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_toolchest}
\alias{install_toolchest}
\title{Toolchest Installation}
\usage{
install_toolchest()
}
\description{
Installs and loads the \code{toolchest_client} Python package.
}
\note{
The Python package interface is stored in the global variable \code{toolchest_client}.
}
