% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken2.R
\name{kraken2}
\alias{kraken2}
\title{Kraken 2 Client}
\usage{
kraken2(
  kraken2_args = "",
  input_path = NULL,
  output_path = NULL,
  input_name = "input.fastq",
  output_name = "output.fastq"
)
}
\arguments{
\item{kraken2_args}{Custom arguments given to function.}

\item{input_path}{Local path of input file (to be uploaded to AWS).}

\item{output_path}{Local path of output file (to be downloaded from AWS).}

\item{input_name}{Name of input file inside input AWS S3 bucket.}

\item{output_name}{Name of output file to be uploaded to output AWS S3 bucket.}
}
\description{
Starts a query for Kraken 2 using Toolchest.
}
\details{
If left unspecified, input_path and output_path can be selected by the user
manually while running the kraken2() function.
}
