% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{kraken2}
\alias{kraken2}
\title{Kraken 2 Client}
\usage{
kraken2(
  tool_args = "",
  inputs = NULL,
  output_path = NULL,
  database_name = "standard",
  database_version = "1"
)
}
\arguments{
\item{tool_args}{(optional) Additional arguments to be passed to Kraken 2.}

\item{inputs}{Path (client-side) to be passed in as input.}

\item{output_path}{Path (client-side) where the output will be downloaded.}

\item{database_name}{(optional) Name (string) of database to use for Kraken 2 alignment.}

\item{database_version}{(optional) Version (string) of database to use for Kraken 2 alignment.}
}
\description{
Starts a query for Kraken 2 using Toolchest.
}
\details{
Currently, only a single file is supported as input.

If left unspecified, inputs and output_path can be selected by the user
manually.
}
\note{
If the database is unspecified, Kraken 2 will use the standard
database, using index files generated by Ben Langmead's lab at
\url{https://benlangmead.github.io/aws-indexes/k2}.
}
