% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{kraken2}
\alias{kraken2}
\title{Kraken 2 Client}
\usage{
kraken2(inputs = NULL, output_path = NULL, tool_args = "")
}
\arguments{
\item{inputs}{Path or list of paths (client-side) to be passed in as input.}

\item{output_path}{Path (client-side) where the output will be downloaded.}

\item{tool_args}{(optional) Additional arguments to be passed to Kraken 2.}
}
\description{
Starts a query for Kraken 2 using Toolchest.
}
\details{
(Currently, only single .fastq inputs are supported.)

If left unspecified, inputs and output_path can be selected by the user
manually.
}
