% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{kraken2}
\alias{kraken2}
\title{Kraken 2 Client}
\usage{
kraken2(
  tool_args = "",
  read_one,
  read_two = NULL,
  output_path = NULL,
  database_name = "standard",
  database_version = "1",
  is_async = FALSE
)
}
\arguments{
\item{tool_args}{(optional) Additional arguments to be passed to Kraken 2.}

\item{read_one}{Path of input file to be passed in as read one.}

\item{read_two}{(optional) Path of input file to be passed in as read two.}

\item{output_path}{(optional) Path to a directory where the output file(s) will be downloaded.}

\item{database_name}{(optional) Name (string) of database to use for Kraken 2 alignment.}

\item{database_version}{(optional) Version (string) of database to use for Kraken 2 alignment.}

\item{is_async}{(optional) Whether to run a job asynchronously. Defaults to false.}
}
\value{
Reference to an object with output location data.
}
\description{
Starts a query for Kraken 2 using Toolchest.
}
\details{
If \code{output_path} is left unspecified, downloading will be skipped.
}
\note{
If the database is unspecified, Kraken 2 will use the standard
database, using index files generated by Ben Langmead's lab at
\url{https://benlangmead.github.io/aws-indexes/k2}.
}
