% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{megahit}
\alias{megahit}
\title{MEGAHIT Client}
\usage{
megahit(
  tool_args = "",
  read_one = NULL,
  read_two = NULL,
  output_path = NULL,
  interleaved = NULL
)
}
\arguments{
\item{tool_args}{(optional) Additional arguments to be passed to MEGAHIT.}

\item{read_one}{(optional) \code{-1} inputs. Path (client-side, string) or list of paths for read 1 of paired-read input files.}

\item{read_two}{(optional) \code{-2} inputs. Path (client-side, string) or list of paths for read 2 of paired-read input files.}

\item{output_path}{Path (client-side) where the output will be downloaded.}

\item{interleaved}{(optional) Path (string) or list of paths for interleaved paired-end files.}
}
\description{
Starts a query for MEGAHIT using Toolchest.
}
\details{
If left unspecified, read_one, read_two, and output_path can be selected by the user
manually.
}
