% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rapsearch2}
\alias{rapsearch2}
\title{RAPSearch2 Client}
\usage{
rapsearch2(tool_args = "", inputs, output_path = NULL)
}
\arguments{
\item{tool_args}{(optional) Additional arguments to be passed to RAPSearch2.}

\item{inputs}{Path to a FASTA/FASTQ file that will be passed in as input.}

\item{output_path}{(optional) Base path to where the output file(s) will be downloaded. (Functions the same way as the "-o" tag for Rapsearch.)}

\item{database_name}{(optional) Name of database to use for RAPSearch2 alignment. Defaults to SeqScreen DB.}

\item{database_version:}{(optional) Version of database to use for RAPSearch2 alignment. Defaults to 1.}
}
\value{
Reference to an object with output location data.
}
\description{
Runs Diamond in BLASTx mode via Toolchest.
}
