% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{STAR}
\alias{STAR}
\title{STAR Client}
\usage{
STAR(
  tool_args = "",
  read_one = NULL,
  read_two = NULL,
  output_path = NULL,
  database_name,
  database_version = NULL
)
}
\arguments{
\item{tool_args}{(optional) Additional arguments to be passed to STAR.}

\item{read_one}{Path of input file to be passed in as Read 1 (-1).}

\item{read_two}{(optional) Path of input file to be passed in as Read 2 (-2).}

\item{output_path}{(optional) Path to a directory where the output file(s) will be downloaded.}

\item{database_name}{Name (string) of database to use for STAR mapping.}

\item{database_version}{(optional) Version (string) of database to use for STAR mapping.}
}
\value{
Reference to an object with output location data.
}
\description{
Starts a query for STAR (mapping) using Toolchest.
}
\details{
If \code{output_path} is left unspecified, downloading will be skipped.
}
\note{
\code{read_one} is required. If paired-end reads are not used, the input
file should be passed in as \code{read_one}.

\code{database_name} should be specified as well.
}
