% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{test}
\alias{test}
\title{Test Pipeline Segment}
\usage{
test(tool_args = "", inputs, output_path = NULL, is_async = FALSE)
}
\arguments{
\item{tool_args}{Additional arguments, present to maintain a consistent interface. This is disregarded.}

\item{inputs}{Path or list of paths (client-side) to be passed in as input.}

\item{output_path}{(optional) Path to a directory where the output file(s) will be downloaded.}

\item{is_async}{(optional) Whether to run a job asynchronously. Defaults to false.}
}
\value{
Reference to an object with output location data.
}
\description{
Run a test pipeline segment via Toolchest. A plain text file containing 'success' is returned."
}
\details{
If \code{output_path} is left unspecified, downloading will be skipped.
}
