% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{unicycler}
\alias{unicycler}
\title{Unicycler Client}
\usage{
unicycler(
  tool_args = "",
  read_one = NULL,
  read_two = NULL,
  long_reads = NULL,
  output_path = NULL,
  ...
)
}
\arguments{
\item{tool_args}{(optional) Additional arguments to be passed to Unicycler.}

\item{read_one}{Path of input file (FASTQ) to be passed in as Read 1 (-1).}

\item{read_two}{Path of input file (FASTQ) to be passed in as Read 2 (-2).}

\item{long_reads}{Path of input file (FASTA) to be passed in as long reads (-l).}

\item{output_path}{(optional) Path to a directory where the output file(s) will be downloaded.}

\item{is_async}{(optional) Whether to run a job asynchronously. Defaults to false.}
}
\value{
Reference to an object with output location data.
}
\description{
Runs Unicycler via Toolchest.
}
\note{
If some but not all of \code{read_one}, \code{read_two}, or \code{long_reads} are
unspecified, Toolchest will assume that they are intended to be blank.
If all are left blank, the user will be prompted to provide input file(s)
and option(s).

At least one input filepath is needed for Toolchest to run Unicycler.

If \code{output_path} is left unspecified, downloading will be skipped.
}
